requirements.txt
flask
mistralai== 0.4.2
python-dotenv
requests
streamlit
gunicorn